+++ src/splash.f90
@@ -56,10 +56,11 @@ program splash
 !             interactive buttons now appear in the plotting window;
 !             cursor movement generates context-dependent help;
 !             cube viz: slice through data using scroll wheel on your mouse
-!     3.4.0   : (04/03/22)
+!     3.4.0   : (24/03/22)
 !             density weighted interpolation now applied automatically to projection
 !             plots of quantities that are not densities;
-!             added flags --codeunits or --code to enforce code units from command line
+!             added flags --codeunits or --code to enforce code units from command line;
+!             successfully parse csv files where some of the fields are character strings
 !     3.3.5   : (01/03/22)
 !             bug fix with disappearing sinks in phantom MPI dumps
 !     3.3.4   : (21/01/22)
@@ -527,7 +528,7 @@ program splash
  logical :: ihavereadfilenames,evsplash,doconvert,useall,iexist,use_360,got_format,do_multiplot
  character(len=120) :: string
  character(len=12)  :: convertformat
- character(len=*), parameter :: version = 'v4.0.0 [4th March 2022]'
+ character(len=*), parameter :: version = 'v4.0.0 [24th March 2022]'
 
  !
  ! initialise some basic code variables